+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/util/color/ColorPattern.java
@@ -7,7 +7,7 @@ public interface ColorPattern {
 
     /**
      * Applies this pattern to the provided String.
-     * Output might me the same as the input if this pattern is not present.
+     * Output might be the same as the input if this pattern is not present.
      *
      * @param string The String to which this pattern should be applied to
      * @return The new String with applied pattern
