--- buildsystem-core/src/main/java/de/eintosti/buildsystem/util/color/impl/SolidPattern.java
@@ -1,30 +1,0 @@
-package de.eintosti.buildsystem.util.color.impl;
-
-import de.eintosti.buildsystem.util.color.IlleniumColorAPI;
-import de.eintosti.buildsystem.util.color.ColorPattern;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-public class SolidPattern implements ColorPattern {
-
-    Pattern pattern = Pattern.compile("<SOLID:([0-9A-Fa-f]{6})>|#\\{([0-9A-Fa-f]{6})}");
-
-    /**
-     * Applies a solid RGB color to the provided String.
-     * Output might me the same as the input if this pattern is not present.
-     *
-     * @param string The String to which this pattern should be applied to
-     * @return The new String with applied pattern
-     */
-    public String process(String string) {
-        Matcher matcher = pattern.matcher(string);
-        while (matcher.find()) {
-            String color = matcher.group(1);
-            if (color == null) color = matcher.group(2);
-
-            string = string.replace(matcher.group(), IlleniumColorAPI.getColor(color) + "");
-        }
-        return string;
-    }
-}
