+++ buildsystem-core/src/main/java/de/eintosti/buildsystem/util/color/patterns/GradientPattern.java
@@ -1,0 +1,28 @@
+package de.eintosti.buildsystem.util.color.patterns;
+
+import de.eintosti.buildsystem.util.color.ColorAPI;
+import de.eintosti.buildsystem.util.color.ColorPattern;
+
+import java.awt.*;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class GradientPattern implements ColorPattern {
+
+    private static final Pattern PATTERN = Pattern.compile("<GRADIENT:([0-9A-Fa-f]{6})>(.*?)</GRADIENT:([0-9A-Fa-f]{6})>");
+
+    @Override
+    public String process(String string) {
+        Matcher matcher = PATTERN.matcher(string);
+        while (matcher.find()) {
+            String start = matcher.group(1);
+            String end = matcher.group(3);
+            String content = matcher.group(2);
+            string = string.replace(
+                    matcher.group(), ColorAPI.color(content, new Color(Integer.parseInt(start, 16)),
+                            new Color(Integer.parseInt(end, 16))
+                    ));
+        }
+        return string;
+    }
+}
